mcp_server/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notion-mcp-server"
version = "1.3.0"
description = "Production-hardened Notion MCP server with RAG capabilities"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
redis = ["redis>=5.0.0"]
dev = [
    "pytest>=7.4.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["mcp_server*"]

[tool.black]
line-length = 88
target-version = ['py311']
